pyproject.toml
[project]
name = "ride"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.115.12",
    "geopy>=2.4.1",
    "googlemaps>=4.10.0",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "polyline>=2.0.2",
    "psycopg2>=2.9.10",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.6.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.40",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
